+++ _testutils/thrift_arith_service-consts.go
@@ -5,10 +5,11 @@ package testutils
 
 import (
 	"bytes"
-	"reflect"
 	"context"
 	"fmt"
-	"git.apache.org/thrift.git/lib/go/thrift"
+	"reflect"
+
+	"github.com/apache/thrift/lib/go/thrift"
 )
 
 // (needed to ensure safety because of naive import list construction.)
@@ -18,7 +19,5 @@ var _ = context.Background
 var _ = reflect.DeepEqual
 var _ = bytes.Equal
 
-
 func init() {
 }
-