+++ client/client.go
@@ -13,10 +13,12 @@ import (
 	"sync"
 	"time"
 
+	opentracing "github.com/opentracing/opentracing-go"
 	circuit "github.com/rubyist/circuitbreaker"
 	"github.com/smallnest/rpcx/log"
 	"github.com/smallnest/rpcx/protocol"
 	"github.com/smallnest/rpcx/share"
+	"go.opencensus.io/trace"
 )
 
 const (
@@ -208,6 +210,15 @@ func (client *Client) Go(ctx context.Context, servicePath, serviceMethod string,
 	if meta != nil { //copy meta in context to meta in requests
 		call.Metadata = meta.(map[string]string)
 	}
+
+	if _, ok := ctx.(*share.Context); !ok {
+		ctx = share.NewContext(ctx)
+	}
+
+	// TODO: should implement as plugin
+	client.injectOpenTracingSpan(ctx, call)
+	client.injectOpenCensusSpan(ctx, call)
+
 	call.Args = args
 	call.Reply = reply
 	if done == nil {
@@ -224,6 +235,59 @@ func (client *Client) Go(ctx context.Context, servicePath, serviceMethod string,
 	call.Done = done
 	client.send(ctx, call)
 	return call
+}
+
+func (client *Client) injectOpenTracingSpan(ctx context.Context, call *Call) {
+	var rpcxContext *share.Context
+	var ok bool
+	if rpcxContext, ok = ctx.(*share.Context); !ok {
+		return
+	}
+	sp := rpcxContext.Value(share.OpentracingSpanClientKey)
+	if sp == nil { // have not config opentracing plugin
+		return
+	}
+
+	span := sp.(opentracing.Span)
+	if call.Metadata == nil {
+		call.Metadata = make(map[string]string)
+	}
+	meta := call.Metadata
+
+	err := opentracing.GlobalTracer().Inject(
+		span.Context(),
+		opentracing.TextMap,
+		opentracing.TextMapCarrier(meta))
+	if err != nil {
+		log.Errorf("failed to inject span: %v", err)
+	}
+}
+
+func (client *Client) injectOpenCensusSpan(ctx context.Context, call *Call) {
+	var rpcxContext *share.Context
+	var ok bool
+	if rpcxContext, ok = ctx.(*share.Context); !ok {
+		return
+	}
+	sp := rpcxContext.Value(share.OpencensusSpanClientKey)
+	if sp == nil { // have not config opencensus plugin
+		return
+	}
+
+	span := sp.(*trace.Span)
+	if span == nil {
+		return
+	}
+	if call.Metadata == nil {
+		call.Metadata = make(map[string]string)
+	}
+	meta := call.Metadata
+
+	spanContext := span.SpanContext()
+	scData := make([]byte, 24)
+	copy(scData[:16], spanContext.TraceID[:])
+	copy(scData[16:24], spanContext.SpanID[:])
+	meta[share.OpencensusSpanRequestKey] = string(scData)
 }
 
 // Call invokes the named function, waits for it to complete, and returns its error status.
@@ -456,6 +520,9 @@ func (client *Client) send(ctx context.Context, call *Call) {
 		req.Payload = data
 	}
 
+	if client.Plugins != nil {
+		client.Plugins.DoClientBeforeEncode(req)
+	}
 	data := req.Encode()
 
 	_, err := client.Conn.Write(data)
@@ -500,11 +567,13 @@ func (client *Client) input() {
 		}
 
 		err = res.Decode(client.r)
-		//res, err = protocol.Read(client.r)
-
 		if err != nil {
 			break
 		}
+		if client.Plugins != nil {
+			client.Plugins.DoClientAfterDecode(res)
+		}
+
 		seq := res.Seq()
 		var call *Call
 		isServerMessage := (res.MessageType() == protocol.Request && !res.IsHeartbeat() && res.IsOneway())
@@ -593,8 +662,8 @@ func (client *Client) input() {
 			client.Plugins.DoClientConnectionClose(client.Conn)
 		}
 		client.pluginClosed = true
-		client.Conn.Close()
 	}
+	client.Conn.Close()
 	client.shutdown = true
 	closing := client.closing
 	if err == io.EOF {
@@ -638,6 +707,7 @@ func (client *Client) heartbeat() {
 
 	for range t.C {
 		if client.shutdown || client.closing {
+			t.Stop()
 			return
 		}