+++ client/plugin.go
@@ -3,6 +3,8 @@ package client
 import (
 	"context"
 	"net"
+
+	"github.com/smallnest/rpcx/protocol"
 )
 
 // pluginContainer implements PluginContainer interface.
@@ -99,6 +101,34 @@ func (p *pluginContainer) DoClientConnectionClose(conn net.Conn) bool {
 	return true
 }
 
+// DoClientBeforeEncode is called when requests are encoded and sent.
+func (p *pluginContainer) DoClientBeforeEncode(req *protocol.Message) bool {
+	var handleOk bool
+	for i := range p.plugins {
+		if plugin, ok := p.plugins[i].(ClientBeforeEncodePlugin); ok {
+			handleOk = plugin.ClientBeforeEncode(req)
+			if !handleOk {
+				return false
+			}
+		}
+	}
+	return true
+}
+
+// DoClientBeforeEncode is called when requests are encoded and sent.
+func (p *pluginContainer) DoClientAfterDecode(req *protocol.Message) bool {
+	var handleOk bool
+	for i := range p.plugins {
+		if plugin, ok := p.plugins[i].(ClientAfterDecodePlugin); ok {
+			handleOk = plugin.ClientAfterDecode(req)
+			if !handleOk {
+				return false
+			}
+		}
+	}
+	return true
+}
+
 type (
 	// PreCallPlugin is invoked before the client calls a server.
 	PreCallPlugin interface {
@@ -120,6 +150,16 @@ type (
 		ClientConnectionClose(net.Conn) bool
 	}
 
+	// ClientBeforeEncodePlugin is invoked when the message is encoded and sent.
+	ClientBeforeEncodePlugin interface {
+		ClientBeforeEncode(*protocol.Message) bool
+	}
+
+	// ClientAfterDecodePlugin is invoked when the message is decoded.
+	ClientAfterDecodePlugin interface {
+		ClientAfterDecode(*protocol.Message) bool
+	}
+
 	//PluginContainer represents a plugin container that defines all methods to manage plugins.
 	//And it also defines all extension points.
 	PluginContainer interface {
@@ -132,5 +172,8 @@ type (
 
 		DoPreCall(ctx context.Context, servicePath, serviceMethod string, args interface{}) error
 		DoPostCall(ctx context.Context, servicePath, serviceMethod string, args interface{}, reply interface{}, err error) error
+
+		DoClientBeforeEncode(*protocol.Message) bool
+		DoClientAfterDecode(*protocol.Message) bool
 	}
 )