+++ log/logger.go
@@ -9,7 +9,7 @@ const (
 	calldepth = 3
 )
 
-var l Logger = &defaultLogger{log.New(os.Stderr, "", log.LstdFlags|log.Lshortfile)}
+var l Logger = &defaultLogger{log.New(os.Stdout, "", log.LstdFlags|log.Lshortfile)}
 
 type Logger interface {
 	Debug(v ...interface{})