+++ share/share.go
@@ -11,6 +11,18 @@ const (
 
 	// AuthKey is used in metadata.
 	AuthKey = "__AUTH"
+
+	// OpentracingSpanServerKey key in service context
+	OpentracingSpanServerKey = "opentracing_span_server_key"
+	// OpentracingSpanClientKey key in client context
+	OpentracingSpanClientKey = "opentracing_span_client_key"
+
+	// OpencensusSpanServerKey key in service context
+	OpencensusSpanServerKey = "opencensus_span_server_key"
+	// OpencensusSpanClientKey key in client context
+	OpencensusSpanClientKey = "opencensus_span_client_key"
+	// OpencensusSpanRequestKey span key in request meta
+	OpencensusSpanRequestKey = "opencensus_span_request_key"
 )
 
 var (